tools/trace: don't raise an exception in a ctype callback
To exit the tool when the maximal number of event is reached (-M
option), the tool currently call exit(), which raise a SystemExit
exception. The handling of exception from ctype callback doesn't seem
straightforward and dependent on python version.

This patch avoid the issue altogether by using a global variable
instead.

Closes #3049

Signed-off-by: Jerome Marchand <jmarchan@redhat.com>