libbpf-tools/filelife: Fix error tracking caused by failed unlink
When we try to delete a file with a normal user after creating a file as
root (sudo), the deletion of the file fails, but the kprobe:vfs_unlink is
triggered, which leads to incorrect file lifecycle tracking.

This commit makes the same changes [1] as tools/filelife.py, or you can
see bcc [2] commit 076ccf079dbb ("tools/filelife: Fix error tracking caused
by failed unlink").

[1] https://github.com/iovisor/bcc/pull/4791
[2] https://github.com/iovisor/bcc/commit/076ccf079dbbbd505cd4f6fa511e67f5f1ee9058

Signed-off-by: Rong Tao <rongtao@cestc.cn>