Mark slabratetop.py not supported for >= 6.8 kernel
slabratetop.py failed with 6.8 kernel with

      /virtual/main.c:100:10: fatal error: 'linux/slub_def.h' file not found
        100 | #include <linux/slub_def.h>
            |          ^~~~~~~~~~~~~~~~~~
      1 error generated.

Actually 'kmem_cache' is also not available to bcc any more.
Either copying 'kmem_cache' or using mini-CORE like logic
are too complex and error prone ([1]). So let us mark
the tool is not supported for >= 6.8 kernel and suggest
to use libbpf-tool version of slabratetop if needed.

  [1]. https://github.com/iovisor/bcc/pull/4940

Signed-off-by: Yonghong Song <yonghong.song@linux.dev>