build: remove llvm-dev dependency from libbcc
In #4737 I added depedency on llvm-config.h in order to discover what LLVM version
is being used. This was only meant to be used at build time.
By adding llvm-config.h in bpf_module.h, a public header of libbcc, I introduced
a dependency on llvm-dev to libbcc which is unnecessary.

This change removes the include from `bpf_module.h`, which should be enough, and
add it to `bpf_module.cc` instead.

Fixes #4997

Signed-off-by: Manu Bretelle <chantr4@gmail.com>