blkalgn: encode granularity in lbs steps
The length was previously encoded in sector steps, but this approach is
inefficient since the LBA format can be larger, and that represents the
smallest granularity for I/O operations.

This update encodes the length using the granularity histogram in LBA
(lbs) steps instead of sector steps. Additionally, the size of the
histogram granularity (hgran) has been increased to accommodate higher
values at very low granularities, such as 16 MiB with a 512-byte step.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>